Organize the repository by roles
Summary:
A role is defined as a set of components to achieve a feature.

A role is so composed of one or several applications, each
deployed by a component.

A component is a collection of deployment script and files to deploy.

This gives the following roles/<role>/<component> tree:

    .
    ├── roles
    │   ├── mailserver
    │   │   └── systemd-unit
    │   │       └── files
    │   └── paas-docker
    │       └── systemd-unit
    │           └── files

Test Plan: no op

Reviewers: Kaliiixx

Differential Revision: http://devcentral.nasqueron.org/D134