Allow to update Mumble server certificates
Summary:
The Mumble server certificate is managed by Let's encrypt.
It's renewed on Ysul and deployed to nginx folder.

We then need to copy new certificates to Mumble jail
and restart the Murmur server.

This code takes care of this copy/restart operation.
A follow-up change should amend Let's encrypt configuration
to call this script automatically when the certificate is renewed.

Ref. T853.

Test Plan:
Tested on Ysul to deploy the current certificate.

Copy, permissions and restart works fine.

Reviewers: Theaptos, Sandlayth

Subscribers: fauve

Differential Revision: https://devcentral.nasqueron.org/D691