Allow to fetch, compile and install an eggdrop automatically
Summary:
Eggdrop users tend to prefer to compile themselves the software
instead to use a packaged one.

But on Debian, GCC must be instructed to use C89 standard, and
on every OS, paths to TCL headers/libs tend to need to be manually set.
See T1033 for some difficulties to compile one on Eglide.

This installer takes care of these details:
Run `install-eggdrop` to get the last version in ~/eggdrop directory.

Thanks to @replicatorbe for the name/scope advice, and to @Sandlayth
for tr suggestion.

Test Plan: Script has been tested in pure sh environment on Eglide and Ysul.

Reviewers: replicatorbe, Sandlayth

Subscribers: Sandlayth, replicatorbe, tomjerr

Maniphest Tasks: T1033

Differential Revision: https://devcentral.nasqueron.org/D695