Claim IPv6 for Odderon
Summary:
To have a knowledge base on Freenode, a darkbot instance is currently
considered. The name will be Odderon, from an hypothetical particle.

See https://en.wikipedia.org/wiki/Pomeron.

The associated vhost is inspired from this article:

The Odderon is the leading exchange in hadronic scattering processes at
high energies in which negative charge conjugation and parity quantum
numbers are transferred in the t-channel. We review the origin of the
Odderon in Regge theory, its status in perturbative and nonperturbative
Quantum Chromodynamics, as well as its phenomenology.
 — Carlo Ewerz, The Odderon in Quantum Chromodynamics

The vhost and associated IP are restricted to the Nasqueron projects,
with a slot open for a new darkbot, for example Pomeron.

Fixes T1128.

Test Plan:
* `vhosts`
* `nslookup -type=AAAA quantum.chromodynamics.at.nasqueron.org`
* `nslookup -type=PTR 2001:470:1f13:896::17`

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1128

Differential Revision: https://devcentral.nasqueron.org/D856