Create nasqueron-irc group on Eglide
Summary:
This group is intended to run IRC bots on Eglide
for Nasqueron services, like Odderon, a darkbot.

When an account is created for an IRC bot, a suoders file can then
be provided to allow bot managers in this nasqueron-irc group to use
sudo from their accounts to run commands on this account.

To provide groups, two operations are so needed:
  - document in GIDs to avoid conflicts
  - add to pillar/users/shellgroups.sls for actual creation

Test Plan:
```
$ ssh eglide id -Gn
dereckson wheel shell nasqueron-irc
```

Reviewers: Sandlayth, dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D858