Clone and compile darkbot on Eglide
Summary:
The darkbot repository contains:
  - source/: the source code
  - dat/: the configuration
  - build/: the built version

As dat/ is handled in Git, we can't clone this repository to be used directly.

So, we're going to clone it and maintain a build in /opt/darkbot.

Then, we'll deploy our own configuration to /opt/odderon.

This change handles the /opt/darkbot clone and compile part.

Ref T1129.

Test Plan: Run `/opt/darkbot/build/darkbot`.

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1129

Differential Revision: https://devcentral.nasqueron.org/D861