Allow nasqueron-irc to control Odderon service
Summary:
f7788aa69fe6 deployed a systemd unit to control the darkbot.

Before, 618a748f0998 offered to nasqueron-irc group the right
to run any command as odderon user.

This is coherent to allow the same group to stop or restart the bot
through systemd.

Test Plan:
As a member of the nasqueron-irc group:
`sudo systemctl restart odderon`

Reviewers: dereckson

Reviewed By: dereckson

Subscribers: Sandlayth

Differential Revision: https://devcentral.nasqueron.org/D888