Mark explicitly files as deployed by SaltStack
Summary:
Configuration files are added here with the intent they will be
managed in the repository and SaltStack will update them.

As such, we mark explicitly files as auto-generated, so tools
can recognize them, with an explanation where to edit them.

The source file parameter isn't redundant: if you open a file
directly on the target server, it allows to know what files to
edit without hunting.

This repository contains files intended to be managed in the repo.

References for comments syntax:

  - http://patrakov.blogspot.be/2011/01/writing-systemd-service-files.html
  - http://www.rsyslog.com/doc/v8-stable/configuration/basic_structure.html#comments
  - https://access.redhat.com/documentation/en-US/Red_Hat_Enterprise_Linux/6/html/Deployment_Guide/sec-Configuring_Yum_and_Yum_Repositories.html
  - https://www.sudo.ws/man/1.8.13/sudoers.man.html#Other_special_characters_and_reserved_words
  - https://www.gnu.org/software/make/manual/html_node/Makefile-Contents.html
  - https://www.freebsd.org/doc/en_US.ISO8859-1/books/pmake/comments.html
  - http://man.he.net/man5/authorized_keys

Headers have been normalized when missing.

Test Plan: Assert comments are valid for each format

Reviewers: Sandlayth, dereckson

Reviewed By: dereckson

Subscribers: amj

Differential Revision: https://devcentral.nasqueron.org/D891