Assign role webserver-legacy to Ysul
Summary:
The webserver-legacy role contains sites not deployed
through the Docker engine, but HTML/CSS/JS or PHP sites
served directly on the dev/staging server.

As automation is faster than provide a solution to put
these sites in production, we can assign the role to
the relevant server.

Test Plan: no-op, as the role is already applied to Ysul

Reviewers: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1045