Install Varnish on Ysul
Summary:
To deploy a Varnish cache, a specific role to assign
to front-end servers 'webserver-varnish' is provided.

The Varnish package is called Varnish everywhere,
excepted currently on FreeBSD where it's `varnish5`.
As such, this change adds the package name to the map.

Ref. T1243.

Test Plan:
When the Salt ran on Ysul:
`Oct  9 23:52:35 ysul pkg: varnish5-5.2.0 installed`

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1243

Differential Revision: https://devcentral.nasqueron.org/D1046