Publish forest nodes definition
Summary:
A node is defined as a server (could be baremetal, VM, a FreeBSD jail
or even a LXC container). A forest is defined as a set of nodes.

The goal of a forest is to allow to apply a configuration to different
nodes, like populate shell users account to Eglide servers.

The nodes definition file allows to centralize preferences by servers
for networking and roles.

Test Plan: Use the pillar in Salt IPv6 network state (D1073).

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1072