Improve Salt master role
Summary:
Create salt and deploy user/group in addition to the sudo capabilites.
This changes does NOT populate those groups. That will be done in the
forest account provisionning follow-up change.

The new Salt directory structure is given to the salt-wrapper configuration.
We currently let the woodscloud test values used for salt-wrapper development,
but we aren't sure it's still required. If so, it's to move to /opt/salt too.
Fixes T1272.

An init state has been prepared, to allow to only include roles/saltmaster.

Ref T1273

Test Plan: Apply state on Ysul

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1273, T1272

Differential Revision: https://devcentral.nasqueron.org/D1078