Upgrade PHPUnit to PHPUnit 6
Summary:
Primary development is now PHP 7, so we require PHPUnit 6.

This fixes the issue the devel/phpunit port pulled PHP 5.6 on Ysul.

Test Plan:
Ideally, we would like a `phpunit --version` to reply 6.4.3.

Currently, this fails, as the devel/phpunit6 binary package hasn't been
published yet. So alternative test is to install it manuall from ports,
then replay the state, and it's well a no-op.

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1085