Switch sphinx package to use Python 3
Summary:
Currently, it used Python 2, to get more Sphinx packages.

But apparently, we only use the Alabaster theme and nothing fancy,
so we can use the genuine package.

As Nasqueron policy is to use Python 3 software when available,
we switch to the py36- package.

Test Plan: http://docs.nasqueron.org/salt-wrapper/ has been rebuilt with this package.

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1092