Provide services for eggdrop
Summary:
Compatible only with FreeBSD.

The rc service provided has an instances parameter to write all
the eggdrop names, so a unique service can be used for all of them.
The code is based on the MySQL FreeBSD port instances handling.

`service eggdrop status` will get an overview of all the eggdrops
in one command. It also allows to run them at reboot or stop it
as once.

Ref T1309

Test Plan:
  - `service eggdrop status`
  - `service eggdrop status Wearg`
  - TC2 has been launched through `/usr/local/etc/rc.d/eggdrop start TC2`

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1309

Differential Revision: https://devcentral.nasqueron.org/D1236