Create php-fpm pool for api.nasqueron.org/servers-log
Summary:
**Facultative deployment**

It's now possible to define a site in web_php_sites
to create a php-fpm pool, but not to deploy it:
only define 'user' and 'php-fpm' parameters,
but skip 'source' and 'target'.

**Environment**

A pool can define specific environment variables.
The state now reads the 'env' parameter in the
web_php_sites pillar entry.

This adheres to the 12 factor app environment pattern.

Test Plan: Ensure the php-fpm pool for the user exists

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1312