Run a secondary SSH server for OTP purpose
Summary:
On Nasqueron servers, sshd on the port 22 is configured to accept only keys.
That configuration helps the user to know the passphrase prompt
is managed by their SSH client when they don't use an agent.

This situation could become more confusing if we add OTP, an interactive
prompt handled by the server.

To avoid such confusion, we run two SSH servers:
  - on the port 22: public key authentication only
  - on the port 5022: key + OTP

Test Plan: Deploy on Ysul and Eglide

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1336