Don't provide rsyslog default configuration everywhere
Summary:
The rsyslog unit in the core role is targeted to the Debian image
built by Scaleway.

This change introduces a new `fixes` key in the node dictionary,
to list possible fixes to common and long-term problems.

This is a follow-up for ed2a64f0931f.

Test Plan:
`salt '*' state.apply roles/core/rsyslog test=True`

The state correctly applies to Eglide only.

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1372