Install Let's encrypt client as a Docker container
Summary:
The Let's encrypt client certbot is installed according two methods:
  - on regular servers, as a distro/OS package
  - on Docker engines, as a container and a wrapper script

This change takes care of the second method.

The certbot/certbot image is used (previously, the letsencrypt
image on quay.org was used), as officially maintained by EFF.

A `certbot` wrapper command runs the container with the correct
volumes mounted to be able to generate or renew certificates.

Test Plan:
  - `salt equatower state.apply roles/paas-docker/docker/images`
  - `salt equatower state.apply roles/paas-docker/wrappers`
  - `certbot <some commands>`

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1422