Allow a container not to be linked to a nginx configuraiton
Summary:
Change 4aaa13576f3b was based on the fact every Docker container
is associated to a nginx configuration.

This situation was modified with change ae11316531f9, with Jenkins
slaves not matching a web server.

This change fixes the nginx unit to only try to populate a file
when the host parameter is defined in the pillar container entry.

Test Plan: Tested on Equatower

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1477