Use package account instead of to create one
Summary:
The Debian package quassel-core creates a "quasselcore" user
and a "quasselcore" service instead of "quassel", so manual
provisionning was prepared, as the quasselcore stuff weren't
immediately noticed.

This change cleans the user provisionning, as it's easier
to maintain in the long term if we shift the the user/service
responsibility to the distribution package.

This commit partially reverts 00bacbfbc607.

Ref T1384.

Test Plan: Restart Debian service

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1384

Differential Revision: https://devcentral.nasqueron.org/D1530