Generate the FreeBSD packages repository signature fingerprints
Summary:
When pkg repo receives a public key as argument, it doesn't publish
fingerprints to the repository. The client needs to get the public key.

Yet, 643db9983997 picked the fingerprint configuration. This configuration
allows more flexibility (several keys, a trusted and a revoked list), and
as all is already ready to use this mode, this changes applies this mode
to pkg repo too.

A benefit is we can separate signature and package builder responsibities
in the future if such a need arises, for example if we've several builders.

Reference: man 8 pkg-repo

Test Plan:
  - repository test: `sudo -u builder make clean all`
  - pkg client test: `sudo pkg update -f -r Nasqueron`

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1553