Deploy Zed rC hypership content to /srv/zed/content
Summary:
The hypership user-generated content, stories, travel rules
have been extracted to their standalone repository.

Some of this content isn't intended to be served publicly,
like stories (playable through UI and gamestory library)
or travel rules. The user content is intended to be served
as web static content, but as such, it should perhaps use
an independant domain for better security.

So, to maximize flexibility, we deploy this repository
content outside the wwwroot directory, and rely on nginx
to pick what they need.

This operational decision should be also checked against
the current zed code, to be sure the content path can be
configured.

Test Plan: Check /srv/zed/content

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1583