Allow arbitrary images to provision Jenkins slave
Summary:
Jenkins execution nodes can be tagged, so a Jenkins job runs on a server
ready for that.

Instead to create a unique 'jack of all trades' Jenkins images, we want to
be able to create more specialized images, for different needs or set of
configurations.

For example, we need to provision a slave with Rust capability.

Ref T1422
Ref T1406

Test Plan: Check the state is no op as far as Jenkins slaves for CD are concerned.

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1422, T1406

Differential Revision: http://devcentral.nasqueron.org/D1728