Map the Phabricator credentials ID
Summary:
Currently, Phabricator credentials are described by number.
This solution isn't the best for maintenance and doesn't allow secrets
replacement easily (what if we want to hunt the number 2 in the repo?).

This change provides a new pillar entry, zr_credentials, to map the IDs
to dictionary keys.

The Zemke-Rhyne execution module is updated to support keys expressions
(with . or : as separatator) in addition to numeric values.

The key dotted format has been selected to be stable, as the key would
also be a valid expression to fetch the secret in other stores, like Vault.

Ref T1425

Test Plan: `salt equatower zr.get_username nasqueron.cachet.mysql`

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1425

Differential Revision: http://devcentral.nasqueron.org/D1738