Standardize states using pkg.installed everywhere
Summary:
Some entries contained only the state module name pkg
instead of the full function.

This approach was deemed more compatible with requisite logic,
where onyl the state module name and the id are required, without
the function name.

Yet, it's the only case this syntax was used, and as such, it should
be standardized.

If there is an advantage state module is more visible, there is also
a drawback: the function name isn't separated from the parameters.

As such, all states will now use the following syntax:

```
<id>:
   <state module>.<method>:
     - parameter1
     - parameter2
     - …
```

Test Plan: no op

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: http://devcentral.nasqueron.org/D1749