Deploy a MySQL 5.7 container for Bugzilla
Summary:
Bugzilla doesn't currently fully support MySQL 8, see for example
https://github.com/bugzilla/bugzilla/pull/71

As such, the best course of action is to deploy a MySQL 5.7 server.

This change introduces an helper execution module to determine
the exact image:tag string from a default image, and dictionary
arguments. That allows to customize the version of a service.

Ref T205

Test Plan:
Units tests are provided for the execution module.

Deploy the container on Dwellers for the state part.

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T205

Differential Revision: https://devcentral.nasqueron.org/D1855