Ensure a valid id is provided to zr functions
Summary:
**Current behavior**

If an empty string is passed as a zr method, the getcredentials utility is called.
The error message is then put in the template, and error is ignored.

Once Jinja "correctly" succeeded, YAML fails.

**Example*

```
  - DB_USER: A positive integer id is expected.
  - DB_PASSWORD: Usage: getcredentials <id> [password|username|title]    <======================
```

**New behavior**

The zr method won't call the utility without a valid key.

Test Plan: The template rendering nowfails with Pillar key not found ValueError.

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1858