Mount Docker registry API microservice on nginx
Summary:
Pass directly to the Equatower app port the request.

This configuration doesn't require an API entry point, nginx removes
the location part of the request URI. Let's keep it in the pillar to
provide a better view of what serves what, and in the future use it
for api.nasqueron.org vhost configuration.

Create 302 redirections for / to the documentation.

Ref T1461.

Test Plan:
  - https://api.nasqueron.org/docker/registry/repository/getAll
  - https://api.nasqueron.org/docker/registry

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1461

Differential Revision: https://devcentral.nasqueron.org/D1892