Deploy www.espace-win.org with staging area
Summary:
This is a legacy code base started in 2001 and published in 2002.

PHP 7 migration of the main parts is done at www51., but there a probably
a lot of places it's broken (there are calls by reference for example).

So the plan is to serve a 503, and update code base to get a working
archive at www51.

Ref T205.

Test Plan: Browse domains, check redirections.

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T205

Differential Revision: https://devcentral.nasqueron.org/D1929