Allow eggdrops to connect to nickserv at runtime
Summary:
Currently, Freenode is under heavy attack, with a lot of channels
using restricted mode. As such, it's especially important to be
connected to nickserv at startup.

It's possible to ask automatic nickserv login if we set a server
password following the "nickname:password" format.

A difficulty is nickserv passwords is stored in database, and
need the `registry` command to query it. That requires to
reorder options a little bit, and so it introduces a little
less readable configuration, as TCL scripts and settings are
both split in two.

Reference: https://freenode.net/kb/answer/registration#logging-in

Test Plan:
Restart service for Wearg, at login it was connected.

14:18:06 -!-  account  : Wearg

Reviewers: Thibaut120094, dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1931