Fetch and extract released PHP source code
Summary:
A PHP custom build starts with some source code.

When the build is for a released version, it's as simple as
decompress an archive. Version and hash in pillar.

As Jinja doesn't support dictionary expansion, the pure
Python state renderer is used here, to see if it's viable
to avoid the proliferation of read-pillar execution modules.

Ref T1469

Test Plan: `salt-call --local state.sls roles/webserver-legacy/php-builder`

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1469

Differential Revision: https://devcentral.nasqueron.org/D1937