Compile and dump a Pure python state file
Summary:
Salt allows to write formula files in different format,
including pure Python (`#!py`) code.

This code must have a `run` method, which returns a dictionary
of the states to execute.

This script allows to print a YAML representation of such dictionary.

**Known issues**

It only loads the pillar, other __ variables would cause the script to fail.

Test Plan: See D1937 for a sample run.

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1938