Provision redis containers
Summary:
Mastodon, Sentry and Pixelfed require a Redis instance.
For Sentry, it has been configured at d8d30a6b35f4.

This change provisions the container without any effort
to bind a port, as it's intended to be used with a network
or a link.

Like MySQL, it allows to customize the image and version
with the facultative image and version parameters.

Test Plan: Deploy redis container for Sentry on Equatower

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1984