Listen Docker registry port on internal network
Summary:
Docker custom images are built on Dwellers,
but registry is hosted on Equatower.

Plan is to use IntraNought secondary NIC between the servers
for internal communication. Meanwhile, a GRE tunnel has
been prepared between the two servers.

The registry container is reconfigured to listen this IP.

Test Plan: Access to the registry from Dwellers

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2001