Clean up Sentry pillar configuration
Summary:
Both nginx and containers SLS rely on the keys in docker_containers
pillar entry to select what items to provision.

As such, we need to rename one of the service into 'sentry'
(an alternative was to rename files to sentry_web).

This is a follow-up for 52ed226f98f8.

Ref T1360.

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1360

Differential Revision: https://devcentral.nasqueron.org/D2005