Check docker_container app_port/host
Summary:
If an entry under docker_containers has `host` defined, it needs also
to have `app_port` defined. The `host` key is used by nginx to determine
if a web configuration should be generated for this container.

This change offers an unit test to validate the pillar configuration,
and fixes the issue for exim container introduced in 3c52ac6a.

Test Plan: New config test provided

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2007