Integrate legacy OpenSSL 1.0 deployment
Summary:
PHP 5.6 also requires OpenSSL 1.0, but this configuration
was only manually applied on Eglide, not provisionned on Ysul,
as a part of now abandoned D924 testing.

This change reintegrates this configuration to the shellserver role,
and takes care it also works on FreeBSD.

This is a follow-up for 4f5ee490.

Test Plan: `salt-call --local roles/shellserver/userland-software/openssl-legacy`

Reviewers: dereckson

Reviewed By: dereckson

Subscribers: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2086