Configure native IPv6 for FreeBSD servers
Summary:
This change introduces two new settings for the node description in pillar:
  - network > ipv6_native, boolean to indicate a server uses IPv6 from its router
  - network > ipv6_interface, should be the same of ipv4_interface in most cases

On FreeBSD, we're not trying to configure IP, as it's the responsibility of the
router configuration, but instead configure to accept such router advertisement.

The IPv6 should still be documented, especially if we want to generate server
homepages through Salt too.

Fixes T1552.

Test Plan: Test on WindRiver

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1552

Differential Revision: https://devcentral.nasqueron.org/D2111