Don't install WDDX PHP extension anymore
Summary:
The WDDX extension has been moved to PECL with the PHP 7.4 release.

It wasn't a standard extension, and was installed, as a Zed dependency.
As we're refactoring sessions in Zed not to use WDDX format anymore,
there is no need to maintain it.

Reference: https://code.zed.dereckson.be/T103
Reference: https://www.php.net/manual/en/migration74.removed-extensions.php

Test Plan: Deploy on WindRiver

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2122