Sync Snap installation with packages
Summary:
Snap needs a small delay to be ready once the service is enabled
to be able to use it to install a package.

To restart it allows a consistent reproducible success.

Fixes T1568.

Test Plan:
Tested on EquaTower (CentOS) with Starship installation:
`salt equatower state.sls roles/core/userland-software`

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1568

Differential Revision: https://devcentral.nasqueron.org/D2147