Use deploy account to clone wwwroot51
Summary:
/var/51-wwwroot is a kludge to offer web-browsable directories outside
home directories for development purpose with web preview capability.

The owner of the subdirectory is so a developer or a service user.
But as this repository doesn't escrow SSH keys for them, it can't be
cloned using this user account. The deploy account is so used, with
a post-clone chown step.

It's a complicated Salt syntax, and perhaps it would be nicer to have
a custom state to do those operations in the right order.

Ref T1559

Test Plan: Test on WindRiver

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1559

Differential Revision: https://devcentral.nasqueron.org/D2154