Generate Nano configuration
Summary:
Read files in nano share directory, and autogenerate a configuration
to include them all.

Fixes T1589.

Test Plan:
Unit tests for the execution module.

Runs on WindRiver for the state module and the state config.

Flake8 reports __salt__ and __opts__ doesn't exist, so the
linter has been configured to be lenient with this kind of reports.

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1589

Differential Revision: https://devcentral.nasqueron.org/D2202