Ignore containers if zr isn't available
Summary:
To render the Jinja templates, we need zr. If the states can revover
from something not available (Docker isn't avaible for example when
we run this role for the first time), to produce states, we need the
templates to be rendered.

This creates a non ideal two pass situation, pending Vault setup.

Ref T1595.

Test Plan: Tested on Dwelles

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1595

Differential Revision: https://devcentral.nasqueron.org/D2212