Use net-tools to setup Dwellers IPv6 tunnel
Summary:
As documented at https://agora.nasqueron.org/Dwellers, at some point
the ip tunnel method stopped to work, but with ifconfig, we still can
configure it.

As it's reproducible on a fresh CentOS 8.1 installation, this change
updates the script with ifconfig instructions and ensure net-tools is
installed (that's the package providing ifconfig).

Ref T1595.

Test Plan: Tested on Dwellers, ping from and to Ysul

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1595

Differential Revision: https://devcentral.nasqueron.org/D2216