Provide PaaS Docker nginx base content
Summary:
If most of the requests are served as proxy_pass
to Docker containers back-end, two "base" cases
need also to be handled:

  - a request to the server domain
  - a request to a domain not served

The content for server homepage should be taken
from the server-homepages repository, but there
isn't currently an API to fetch it or a script
to build it. Instead, we provide a temporary
file Salt won't try to replace afterwards.

If a domain isn't served, we need to provide
information about possible and reasonnable causes,
and links for assistance.

Ref T1595

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1595

Differential Revision: https://devcentral.nasqueron.org/D2228