Drop cvs2svn
Summary:
Initially introduced in 690ca84c957b as requested on T111.

This utility hasn't been updated against recent Subversion code base
and so on FreeBSD requires subversion-lts instead of subversion,
creating a conflict with other packages.

CVS repositories are becoming rare enough to justify install it only
when needed. So the easiest is to drop it.

Test Plan: Manually remove it and update packages with pkg

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2295