Refactor network unit
Summary:
With GRE tunnels addition, we start to have a long unit, especially
as it needs to provide solutions for several OS and distribution,
even if currently it doesn't try to configure IPv4 for Linux.

We sort the files/ by topics and ensure each file explictly says
the protocol it configures.

This change is no-op for IPv4, GRE, IPv6 interfaces and tunnels.

This change makes mandatory instead of facultative the IPv6 routing
for FreeBSD.

Test Plan: Deploy on WindRiver

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2307