Install OpenSearch
Summary:
OpenSearch is the AWS fork of ElasticSearch and Elastic stack.

This role downloads software and configure OpenSearch as a single-node
cluster. It contains all the bricks to configure a multi-node cluster.

Deploy a single-node cluster "infra-logs" on CloudHugger to validate the role.

Ref T1624.

Test Plan: `es-query _cluster/health?pretty`

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1624

Differential Revision: https://devcentral.nasqueron.org/D2449