Tweak OpenSearch installation
Summary:
Order of operations matters: to initialize the security plugin,
we need OpenSearch to be running.

Don't try to be smart extracting the archive, we can't recover
from /opt deletion if we do that.

This s a follow-up for 04e70ee4e1b33c7.

Ref T1624

Test Plan: Redeploy to CloudHugger

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1624

Differential Revision: https://devcentral.nasqueron.org/D2451