Enable oidentd service
Summary:
On WindRiver, oidentd is installed but service is disabled:

```
$ /usr/local/etc/rc.d/oidentd status
Cannot 'status' oidentd. Set oidentd_enable to YES in /etc/rc.conf
or use 'onestatus' instead of 'status'.

$ /usr/local/etc/rc.d/oidentd onestatus
oidentd is not running.
```

This change enables the service.

Ref T1683.

Test Plan: Check service is enabled and running on WindRiver.

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1683

Differential Revision: https://devcentral.nasqueron.org/D2540