Compute CIDR prefix from netmask for RHEL network scripts
Summary:
The network scripts use a CIDR notation like `PREFIX=24`
instead of a netmask notation.

Salt offers a cidr_to_ipv4_netmask but not the other way
around, so a new network execution module is offered.

This is a follow-up for f9dbdab3b579.

Test Plan: Unit tests are provided for network module

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2568