Simplify network configuration
Summary:
The network unit wants to configure "IPv4", "IPv6" and "private interface".

The private interface configuration is an IPv4, so there is redundant code
to maintain there.

This change introduces an interface dictionary in the nodes pillar where
the ipv4/ipv6/private configuration is migrated.

MOTD wants a public IPV4 and a gateway, so the node.resolve_network()
parses our configuration to provide that and return private IPv4 for
servers not assigned a public address. The first interface wins.

Elsewhere, configuration won't be changed as the grain ipv4 is used.

Test Plan: Apply to Dwellers and router-001

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2581